refactor: 졸업 계산 응답 결과 디렉터리 위치 수정
