+++ src/main/java/com/plzgraduate/myongjigraduatebe/graduation/application/port/in/response/GraduationResponse.java
@@ -1,4 +1,4 @@
-package com.plzgraduate.myongjigraduatebe.graduation.adpater.in.web.response;
+package com.plzgraduate.myongjigraduatebe.graduation.application.port.in.response;
 
 import static com.plzgraduate.myongjigraduatebe.graduation.domain.model.GraduationCategory.*;
